Fix printf typecast problem
See http://bugzilla.pokylinux.org/show_bug.cgi?id=1473

Without this fix the printed TLS(...) information is incorrect, which
causes the core prelinker code to incorrectly prelink.  This bug triggers
primarily on x86 (32-bit) hosts, targeting x86 (32-bit) targets and arm
targets.

2011-09-13  Mark Hatle <mark.hatle@windriver.com>
        * Fix printf problem causing prelink-rtld issues on x86 (32-bit)
          and arm

+++ trunk/src/rtld/rtld.c
@@ -1120,7 +1120,7 @@ process_one_dso (DSO *dso, int host_paths)
 		   (size_t) (l->l_map_start - cur_dso_ent->dso->base));
 
 	   if (l->l_tls_modid)
-	     printf (" TLS(0x%Zx, 0x%0*Zx)\n", l->l_tls_modid,
+	     printf (" TLS(0x%Zx, 0x%0*Zx)\n", (size_t) l->l_tls_modid,
 		     (int) size_pointer,
 		     (size_t) l->l_tls_offset);
 	   else
